backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-arango==7.9.1
psycopg[binary]==3.1.18
SQLAlchemy==2.0.36
alembic==1.13.2
python-dateutil==2.9.0.post0
google-cloud-resource-manager==1.12.3
google-auth==2.23.4
